+++ src/index.ts
@@ -3,5 +3,5 @@ import { ExtensionContext } from "@foxglove/studio";
 import { initJoyPanel } from "./JoyPanel";
 
 export function activate(extensionContext: ExtensionContext): void {
-  extensionContext.registerPanel({ name: "Joystick", initPanel: initJoyPanel });
+  extensionContext.registerPanel({ name: "POLARIS Joystick", initPanel: initJoyPanel });
 }